+++ edcloud/__init__.py
@@ -3,3 +3,5 @@
 from importlib.metadata import version
 
 __version__ = version("edcloud")
+
+__all__ = ["__version__"]